pykey/requirements.txt
Flask==3.0.3
python-keycloak==4.3.0
requests==2.32.3

